achieve hidding course

+++ resources/views/rooms/index.blade.php
@@ -12,20 +12,23 @@
             <strong>{{ $messageDelete }}</strong>
         </div>
         @endif
-        <div class="lead">
-            <a href="{{ route('rooms.create') }}" class="btn btn-success btn-sm float-right">Add new room</a>
-        </div>
-
+        @if(auth()->user()->id==1)
+            <div class="lead">
+                <a href="{{ route('rooms.create') }}" class="btn btn-warning float-right mb-4">Add new room</a>
+            </div>
+        @endif
         <div class="mt-2">
             @include('layouts.partials.messages')
         </div>
-
-        <table class="table table-striped">
+    @if(count($rooms))
+        <table class="table table-warning">
             <thead>
             <tr>
-                <th scope="col" width="45%">room name</th>
-                <th scope="col" width="45%">capacity</th>
-                <th scope="col" width="5%">Actions</th>
+                <th scope="col" width="20%">room name</th>
+                <th scope="col" width="20%">capacity</th>
+                <th scope="col" width="20%">location</th>
+                <th scope="col" width="30%">notes</th>
+                @if(auth()->user()->id==1)<th scope="col" width="10%">Actions</th>@endif
             </tr>
             </thead>
             <tbody>
@@ -33,15 +36,29 @@
                         <tr>
                             <td>{{ $room->room_name }}</td>
                             <td>{{ $room->capacity }}</td>
-                            <td><a href="{{ route('rooms.edit', $room->id) }}" class="btn btn-info btn-sm">Edit</a></td>
-                            <td>
-                                {!! Form::open(['method' => 'DELETE','route' => ['rooms.destroy', $room->id],'style'=>'display:inline']) !!}
-                                {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                                {!! Form::close() !!}
-                            </td>
+                            <td>{{ $room->location }}</td>
+                            <td>{{ $room->notes }}</td>
+                            @if(auth()->user()->id==1)
+                                <td style="display:flex;">
+                                        <a href="{{ route('rooms.edit', $room->id) }}" class="btn btn-info btn-sm me-2 btn-close-white">Edit</a>
+                                        {!! Form::open(['method' => 'DELETE','route' => ['rooms.destroy', $room->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
+                                        {!! Form::close() !!}
+                                </td>
+                            @endif
                         </tr>
                     @endforeach
             </tbody>
         </table>
+        @else
+        <div class="alert text-black alert-success" role="alert" style="margin-top: 20px;">
+            <h4 class="alert-heading">Sorry<h4>
+            <p>There are not any room yet .</p>
+            <hr>
+            <p class="mb-0">Whenever you need to add a new room, click the yellow button .</p>
+           <h1><a href="{{url()->previous()}}" class="btn btn-secondary"> Back</a></h1>
+           {{-- problem in back --}}
+        </div>
+      @endif
     </div>
 @endsection
