+++ resources/views/users/index.blade.php
@@ -9,23 +9,28 @@
             </div>
         </h1>
         <div class="lead">
-            <a href="{{ route('users.create') }}" class="btn btn-success btn-sm float-right">Add new user</a>
+            @if(auth()->user()->id==1)
+                <a href="{{ route('users.create') }}" class="btn btn-warning float-right mb-4">Add new user</a>
+            @endif
         </div>
 
         <div class="mt-2">
             @include('layouts.partials.messages')
         </div>
         {{-- @livewire('search') --}}
-        <table class="table table-striped">
+        @if(count($users))
+        <table class="table table-warning">
             <thead>
             <tr>
                 <th scope="col" width="5%">#</th>
                 <th scope="col" width="15%">Email</th>
                 <th scope="col" width="15%">Username</th>
                 <th scope="col" width="15%">Role</th>
-                <th scope="col" width="20%">number observation</th>
-                <th scope="col" width="20%">current number_of_observation</th>
-                <th scope="col" width="10%" colspan="3">Actions</th>
+                @if(auth()->user()->id==1)
+                    <th scope="col" width="20%">number observation</th>
+                    <th scope="col" width="15%">current number_of_observation</th>
+                @endif
+                <th scope="col" width="15%" colspan="3">Actions</th>
             </tr>
             </thead>
             <tbody id="user-list" name="users-list">
@@ -37,46 +42,60 @@
                             <td>
                                 <span class="badge bg-danger">{{$user->role}}</span>
                             </td>
-                            <td>
-                                <span class="badge bg-secondary">{{$user->number_of_observation}}</span>
-                            </td>
-                            <td>
-                                
-                                @php
-                                    $current_observations_for_all_users=App\Models\User::with('courses')->whereHas('courses',function($query) use($user) {
-                                        $query->where('user_id',$user->id);
-                                    })->get();
-                                    $dates_distinct=[];
-                                    $times_distinct=[];
-                                @endphp
+                            @if(auth()->user()->id==1)
+                                <td>
+                                    <span class="badge bg-secondary">{{$user->number_of_observation}}</span>
+                                </td>
+                                <td>
+                                    
+                                    @php
+                                        $current_observations_for_all_users=App\Models\User::with('courses')->whereHas('courses',function($query) use($user) {
+                                            $query->where('user_id',$user->id);
+                                        })->get();
+                                        $dates_distinct=[];
+                                        $times_distinct=[];
+                                    @endphp
 
-                                @foreach($current_observations_for_all_users as $current_user)
-                                    @foreach($current_user->courses as $course)
-                                         @if( (!in_array($course->pivot->date,$dates_distinct) && !in_array($course->pivot->time,$times_distinct) ) ||
-                                              ( in_array($course->pivot->date,$dates_distinct) && !in_array($course->pivot->time,$times_distinct) ) ||
-                                              (!in_array($course->pivot->date,$dates_distinct) &&  in_array($course->pivot->time,$times_distinct) ) )
-                                                @php 
-                                                    array_push($dates_distinct,$course->pivot->date);
-                                                    array_push($times_distinct,$course->pivot->time); 
-                                                @endphp
-                                        @endif
+                                    @foreach($current_observations_for_all_users as $current_user)
+                                        @foreach($current_user->courses as $course)
+                                            @if( (!in_array($course->pivot->date,$dates_distinct) && !in_array($course->pivot->time,$times_distinct) ) ||
+                                                ( in_array($course->pivot->date,$dates_distinct) && !in_array($course->pivot->time,$times_distinct) ) ||
+                                                (!in_array($course->pivot->date,$dates_distinct) &&  in_array($course->pivot->time,$times_distinct) ) )
+                                                    @php 
+                                                        array_push($dates_distinct,$course->pivot->date);
+                                                        array_push($times_distinct,$course->pivot->time); 
+                                                    @endphp
+                                            @endif
+                                        @endforeach
                                     @endforeach
-                                @endforeach
 
-                                <span class="badge bg-secondary">{{count($dates_distinct)}}</span>
-                            </td>
-                            <td><a href="{{ route('users.observations', $user->id) }}" class="btn btn-success btn-sm">observations</a></td>
+                                    <span class="badge bg-secondary">{{count($dates_distinct)}}</span>
+                                </td>
+                            @endif
+                            @if(auth()->user()->id==1)<td><a href="{{ route('users.observations', $user->id) }}" class="btn btn-info btn-sm me-2 btn-close-white">observations</a></td>@endif
                             <td><a href="{{ route('users.show', $user->id) }}" class="btn btn-warning btn-sm">Show</a></td>
-                            <td><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm">Edit</a></td>
-                            <td>
-                                {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                                {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                                {!! Form::close() !!}
-                            </td>
+                            @if(auth()->user()->id==1)
+                                <td style="display:flex;">
+                                        <a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm me-2" style="height: 32px;">Edit</a>
+                                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                        {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
+                                        {!! Form::close() !!}
+                                </td>
+                            @endif
                         </tr>
                     @endforeach
             </tbody>
         </table>
+        @else
+        <div class="alert text-black alert-success" role="alert" style="margin-top: 20px;">
+            <h4 class="alert-heading">Sorry<h4>
+            <p>The Program has not any user yet .</p>
+            <hr>
+            <p class="mb-0">Whenever you need to add a new user, click the yellow button .</p>
+           <h1><a href="{{url()->previous()}}" class="btn btn-secondary"> Back</a></h1>
+           {{-- problem in back --}}
+        </div>
+      @endif
 {{-- added --}}
         {{-- <div class="modal show" id="formModal" aria-hidden="true"> --}}
             <div class="modal-dialog" style="display:none">
